# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CAURIS DIGITAL — API
  version: '2.0.0'
  description: |
    Documentation OpenAPI des **routes API HTTP réelles** de CAURIS DIGITAL (CDC V2 §6).

    **Périmètre.** Le CRUD complet de l'espace d'administration (startups, articles,
    événements, partenaires, utilisateurs, messages, inscrits) est implémenté via des
    Server Actions Next.js — des mutations serveur typées de bout en bout, invoquées
    directement depuis les formulaires React, sans appel HTTP JSON classique. Ce
    mécanisme n'a pas de représentation OpenAPI significative (pas de méthode HTTP,
    pas d'URL stable, payload sérialisé par le framework). Voir le CDC V2 §6.3.8
    « Correspondance avec l'implémentation réelle » pour la liste des Server Actions
    et leur fichier source.

    Ce document couvre donc uniquement les endpoints qui sont de vraies routes HTTP :
    lecture publique du contenu, formulaires publics (contact, newsletter),
    authentification, et exports CSV admin.

    **Pourquoi ce périmètre reste pertinent.** Une API REST complète redevient
    nécessaire si un consommateur externe apparaît (application mobile V3, §9.5,
    intégration tierce). Ces routes existantes constituent alors une base déjà
    testée (§12.4) à étendre plutôt qu'à reconcevoir.
  contact:
    name: CAURIS DIGITAL
    url: https://caurisdigital.org
servers:
  - url: https://caurisdigital.org
    description: Production
  - url: http://localhost:3000
...
security: []
...
paths:
...
  /api/newsletter/confirm:
    get:
      operationId: confirmNewsletterSubscription
      tags: [Newsletter]
      summary: 'Étape 2/2 — Confirmer l''inscription (lien cliqué depuis l''email)'
      description: |
        Seule étape où le contact est réellement ajouté à l'audience Resend et où
        l'email de bienvenue part. Une réinscription réactive le compte existant
        (RM-N03) plutôt que de dupliquer la ligne en base. Le jeton est auto-porteur
        (HMAC-SHA256, pas de lookup en base pour la validité) et expire après 48h.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
        - name: firstName
          in: query
          schema: { type: string, maxLength: 80 }
      responses:
        '307':
          description: |
            Toujours une redirection : vers `/newsletter/confirmee` en cas de succès,
            ou vers `/newsletter/erreur?raison={manquant|invalide|expire|serveur}` sinon.
          headers:
            Location: { schema: { type: string } }
  /api/newsletter/unsubscribe:
    get:
      operationId: unsubscribeNewsletter
      tags: [Newsletter]
      summary: Désinscription en un clic (variante historique)
      description: |
        Conservée pour les emails de bienvenue déjà envoyés avec un lien pointant
        directement ici. Les nouveaux emails renvoient plutôt vers la page
        `/newsletter/desinscription` (qui appelle POST ci-dessous après un
        formulaire), pour éviter qu'un simple GET automatique (scanner de sécurité
        email, aperçu de lien) ne désinscrive quelqu'un sans action humaine.
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '307':
          description: |
            Redirection vers `/newsletter/desinscrit` en cas de succès, ou vers
            `/newsletter/erreur?raison={manquant|invalide|serveur}` sinon.
          headers:
            Location: { schema: { type: string } }
    post:
      operationId: unsubscribeNewsletterWithReason
      tags: [Newsletter]
      summary: Désinscription avec message optionnel
      description: |
        Appelée par le formulaire de la page `/newsletter/desinscription`. Même
        jeton (usage `unsubscribe`) que la variante GET ci-dessus ; le message,
        s'il est renseigné, est tronqué à 500 caractères et stocké sur l'abonné
        (visible côté admin) pour comprendre les raisons de départ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
                message: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Désinscription effectuée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: authjs.session-token
      description: |
        Cookie de session NextAuth v5 (JWT). Le JWT ne porte qu'un `sessionToken`
        opaque ; la validité réelle (session non révoquée, compte actif, rôle) est
        vérifiée côté serveur en base à chaque requête (voir `src/lib/require-admin.ts`),
        pas uniquement via la signature du JWT.
  responses:
...
    ValidationError:
      description: Corps de requête invalide (champ manquant, format incorrect, etc.).
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    ServerError:
      description: Erreur serveur inattendue.
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...